remake budget and transaction delete alert with the new alert

+++ src/main/java/myApp/controllers/components/BudgetBox.java
@@ -7,6 +7,7 @@ import io.github.palexdev.materialfx.enums.ButtonType;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.control.Alert;
+import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
@@ -81,20 +82,30 @@ public class BudgetBox extends AnchorPane {
     }
 
     private void deleteBudget(ActionEvent actionEvent) {
-        String sql = "DELETE FROM budget WHERE budgetID = ?";
+        ManualAlert confirm = new ManualAlert(Alert.AlertType.CONFIRMATION, "Confirm Deletion",
+                "Are you sure you want to delete this budget?",
+                "This action cannot be revert!");
 
-        try (Connection conn = ConnectionManager.getConnection();
-             PreparedStatement pstmt = conn.prepareStatement(sql)) {
+        confirm.showAndWait().ifPresent(buttonType -> {
+            if (buttonType.equals(javafx.scene.control.ButtonType.YES)) {
+                String sql = "DELETE FROM budget WHERE budgetID = ?";
 
-            pstmt.setInt(1, budget.getId()); // Use ID here for the operation
-            int affectedRows = pstmt.executeUpdate();
-            if (affectedRows > 0) {
-                this.setVisible(false);
-                this.setManaged(false);
+                try (Connection conn = ConnectionManager.getConnection();
+                     PreparedStatement pstmt = conn.prepareStatement(sql)) {
+
+                    pstmt.setInt(1, budget.getId()); // Use ID here for the operation
+                    int affectedRows = pstmt.executeUpdate();
+                    if (affectedRows > 0) {
+                        this.setVisible(false);
+                        this.setManaged(false);
+                    }
+                } catch (SQLException e) {
+                    showAlert("Error", "Failed to delete budget: " + e.getMessage());
+                }
             }
-        } catch (SQLException e) {
-            showAlert("Error", "Failed to delete budget: " + e.getMessage());
-        }
+        });
+
+
     }
 
     private void updateBudget(ActionEvent actionEvent) {
